requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16.0,<4.0
djangorestframework-simplejwt>=5.4.0,<6.0
psycopg[binary]>=3.2.0,<4.0
Pillow>=10.0.0,<12.0
python-dotenv>=1.0.1,<2.0
reportlab>=4.2.0,<5.0
google-auth>=2.30.0,<3.0
requests>=2.31.0,<3.0
